Orientation matrices functions, allow to manipulate image space
xASL_im_DecomposeAffineTransformation -> allows to define the separate transformations from which a certain transformation matrix is made out, extension of spm_matrix and/or spm_imatrix

xASL_im_DummyOrientationNIfTI -> takes the previous transformation matrix components, and allows to choose which ones to apply to a source NIfTI, to create a dummy NIfTI with the desired orientation matrix. This can then be used to transform the source NIfTI with xASL_spm_reslice into the desired space/orientation.

xASL_im_TestDummyNIfTIOrientation -> allows to test the above