Update ExploreASL_Import.m
Fix to make ExploreASL_Import deal better with multiple time point/visits. If visits are specified as TokenAlias, they are processed and the NIfTI folders in the analysis folder will get the postfixes (e.g. _1 _2 _3), if these are not specified, or if there is only a single visit specified, this won't be appended.