#717 Major fix for backwards compatibility
The major problem was that xASL_adm_LoadParms just dumps all fields of the JSON into the x struct. This is not really helpful and maybe we should do this in a more controlled way in a future release. For now it should mostly be fixed. We use a standardized script called xASL_io_CheckDeprecatedFieldsX for this now. We use the same script for xASL_io_ReadDataPar and we could also integrate it in the LoadX function, we just have to use OldX.x for x there I think.