#639 xASL_imp_DCM2NII_...: Improve comments, fix minor bug
You have to be a bit careful here. The keyword hadamard is also a matlab function, which is why it somehow worked for you, but certainly not as intended. In addition it was theoretically possible that x.modules.asl.bHadamard was true but there was no SeriesDescription and you would have checked that field later without checking for existence first. There was an easy fix for that though. I tested it based on Amnahs Hadamard-4 and it worked perfectly well. One minor warning because of the TotalAcquiredPairs, but maybe we have to see if that would still appear in current develop. I would merge these rather early and then improve it later on.