#778 xASL_imp_...: Minor optimizations
xASL_imp_DCM2NII_Initialize was a misleading name. We mostly read and initialize imPar here. This is mostly based on the sourceStructure.json. I think as a name xASL_imp_Initialize is a bit more straightforward and shorter. Otherwise you would expect this part to be within the DCM2NII submodule.