#778 Move xASL_init_DataLoading: Fix bug of #769
The problem with the way that Henk fixed #769 is that it now tries to load the dataset even if we only ran one of the first import modules. To fix this, we need to ensure that either a derivatives directory or a legacy dataset exists. I think the current method should be able to capture most use-cases, but we sould really start to stop supporting this outdated usage of ExploreASL, it makes the code super complex.