#834 xASL_adm_CreateFileReport: check required fields
This change was on my to-do for a long time now. Especially while testing the different code bugs, we can see that this is required. The CreateFileReport script is not run within a try-catch statement. This is a main reason why it should never crash. Therefore we need proper checks here.