#834 Testing and comparison revamp
Originally I had 10 separate commits here, but it was getting a bit ugly. I squashed them together. I am mainly revamping xASL_bids_CompareStructures here and I created a unit test for it. There is also an update for the testing scripts related to the flavor testing and I moved a lot of functions to separate scripts.