#799 Major edits for a full import module
A few seconds ago I discussed with Jan what will be necessary to make the xASL_Iterate script work for the import module. We need to build up the subject/session structure before the actual import. This can be done, but will require some changes to ExploreASL_ImportMaster. I added Import to the regexp in xASL_Iteration. I moved a lot of fields to the x struct, otherwise we cant run xASL_module_Import from xASL_Iterate either. We wanted to do this anyway, so this is pretty nice. I added two scripts (DetermineSubjectStructure & BasicParameteChecks). The parameter checks are basically the part that was the IO check of xASL_module_Import before. The DetermineSubjectStructure will be the part that determines the subject/session structure based on the sourcedata (higher priority) or temp data (lower priority).