#799 Initial changes for xASL_Iteration import
I had to move a lot of variables and fields. There were even a lot of bugs related to our build up structure. Before we basically assumed the same number of visits for each subject. Still a lot to do and to optimize, especially Bids2Legacy isnt ready and we have to implement the moving of the log files and the deletion of the temp files.